requirements.txt
django = "*"
django-debug-toolbar = "*"
mysqlclient = "*"